+++ interfaces/Prolog/YAP/yap_cfli.hh
@@ -313,7 +313,7 @@ Prolog_get_atom_name(Prolog_term_ref t, Prolog_atom* ap) {
   The behavior is undefined if \p t is not a Prolog compound term.
 */
 inline int
-Prolog_get_compound_name_arity(Prolog_term_ref t, Prolog_atom* ap, int* ip) {
+Prolog_get_compound_name_arity(Prolog_term_ref t, Prolog_atom* ap, size_t* ip) {
   assert(Prolog_is_compound(t));
   YAP_Functor f = YAP_FunctorOfTerm(t);
   *ap = YAP_NameOfFunctor(f);
